Improved Cloudbase-Init installation logic
Changed the logic, so that the method will retry to install Cloudbase-Init
until succeeds.
After installation, it is checked for validity. If it's an invalid
installation, before retrying, a clean-up is done.
This is a better approach since the installation was sometimes previously
unsuccessful and it caused a scenario to fail.